Added loop for input
I added a loop for the input part of the program. The user has 4
attempts to enter the correct result. If he or she enters the right
result the loop will be left before all attempts are used. After
leaving the loop either because of entering the correct result or
the loop has been finished the correct result will be displayed.